Fix entry count metric for lookup caches
The way the `LookupCache#entryCount()` method was defined, caused implementers to write incorrect implementations.

For example `GuavaLookupCache#entryCount()` returned a constant `Gauge<Long>` on the first call (in the constructor
of `LookupCache`) which was never updated.

Additionally, the `CacheTableEntry` tried to access the entry count metric through an invalid object path which
resulted in the metric always being displayed as "NaN" in the web interface.

Fixes #4540