Add EditTokensPage to Components
If a user with the role admin wants to reach the EditTokensPage
he opens the users list and clicks there on a users Edit Tokens.

If a user without admin role want to edit his tokens he will reach
the site via Edit Profile and there he can click on Edit Tokens.