Make API Tokens configurable in web interface (#4550)
* Correct API Token Permission Handling in Backend

The permission handling for tokens was based only on the
username. So a user could only see his own tokens no matter
what kind of permissions he had. Even the admin which should
have full controll over the access could not see or alter the
API Tokens of diffrent users.

* Remove check based on username of request
* Use the RestPermissions check based on the username
* Give a user default permissions for his own API Tokens.

* Add new new TokenList Component

To list and edit API Tokens we need a new Component which
will receive a list of Tokens as well as the necessary
handlers to create and delete the tokens.

* List tokens and add a Copy to Clipboard Button and Delete
  Button
* A from field to create a new API Token

* Add API Token api endpoints to UserStore

To load, create and delte API Tokens we need to add them to
the UserStore.

* Add functions to user store
* Add API Routes

* Add new EditTokensPage

To display a users tokens we add a new page. That page
will contain a Table to display and delete a users tokens
and a form to create a new token.

Add EditTokensPage to Router.

* Add EditTokensPage to Components

If a user with the role admin wants to reach the EditTokensPage
he opens the users list and clicks there on a users Edit Tokens.

If a user without admin role want to edit his tokens he will reach
the site via Edit Profile and there he can click on Edit Tokens.

* Some improvements to TokenList

* Make Tokens hideable
* Update Tokens when delete or create tokens
* Beautify create form of tokens

* Tests for TokenList

- Add tests for TokenList Component
- Add class to TokenList Component for easier testing

* Review Fixes

- In the backend we should use the inheritate function isPermitted instead of
  reimplementing it
- Fix Exception messages
- Fix linter warnings
- rename props create to onCreate and delete to onDelete to be more consistent
- When creating or deleting a token we should indicate a user that the progress
  did start. This might be necessary if the server is not anserwering fast enough.
- Use ClipboardButton instead of reimplementing him
- Group Buttons with ButtonGroup instead of span
- use form submit instead of the button onClick handler, that way a enter works just
  as good as clicking the button
- move react-dom mock to lib to make it reusable
- comment css so we don't forget why this is there
- Fix test accordingly

* Fixes from Review 2

- pervent user name inspection by testing the permissions before loading
  the user
- Do not use wildcards when importing in java!
- rename function for better understanding
- Remove unused route from Routes.jsx
- Use scope for edit tokens instead of postfixing in the name (edit_tokens)