Display Timestamp in User Timezone (#4521)
* Allow timezone of user to be set null

On the client side we need a way to distinguish if a user
set a timezone or not. If a user wishes to not set a timezone
the browser should deciede the timezone.

Till now the server was setting the timezone to UTC if no
timezone was set. Now if the send timezone is null
we set the timezone to null. Also if a user is requested
we send null as timezone if no timezone is configured, rather
then sending UTC.

* Add tests for MessageTableEntry

* Test if the rendering does work
* Test use of browser timezone

* Change choosing the source of timezone

If we choose the timezone in DateTime we did not use the
browser time as a fallback but the timezone of the user admin
(which must not necessarily be the timezone of the current user).

We now instead try to guess the the timezone of the current user
by asking the browser. The fallback after that is still the
admin user and after that we set it to UTC. In the most cases
the browser should provide a time zone.

We need to do this because in a former change, I changed the behaviour
of the server providing a fallback for the timezone of the current user.
The server now does return null if the user did not set a timezone
and wants to use the timezone of the browser.

* Display timestamp in user choosen timezone

The use would like to the displayed timestamps in his choses
timezone.

* Add function renderForDisplay to replace possiblyHighlight
  where we parse the value if the field name is 'timestamp'
  and change the timestamp to user choosen timezone.
  After that we call possiblyHighlight to keep old functionallity.
* Replace possiblyHighlight with renderForDisplay where it is
  called.

* Stablize test result

The test was failing in diffrent timezone since
the time rendered differed from mine computer :)

* Fixes from review

* Avoid using firstNonNull if we already checked if the first
  value is not null.
* Do not fallback to UTC if the caller does give an invalid
  argument. Just tell him it was wrong.
* rename missed caller of possibleHighlight to renderForDisplay
* Show MessageDetail timestamp with Timezone in which it was
  rendered to clarify for the user that it is a parsed value.

* Further fixes from review

* Use Timestamp to convert the time in users timezone.
* Add a little info icon to the Timestamp which tells
  the user by hover, that this timestamp was altered to
  be rendered in his timezone.
* Fix test accordingly

* Fixes from review

- make the info icon a bit darker for better contrast
- fix import of css