Prevent ldap server password leak (#4526)
* Prevent ldap server password leak

The ldap configuration page was setting the input value of the
password field to the actuall password. This way a attacker
could read the ldap password, if he gained admin access to graylog.

This change prevents it by setting the input value of the password
with a placeholder in form of 13 asterisks.

When the gui sends the form field to the server, the server
compares the value with the placeholder and only update the
password if it differs with the placeholder.

* Show empty password field for new ldap config

If there is no ldap password yet set, it should be displayed empty.

* Review fixes

Instead of using the magic string "**********" we send the state
of the password in the GET action. So we tell the GUI if a
password is set or not. On the client side we use that information
to set the placeholder accordingly.

But on the server side we accept the system_password_set flag
so we do not have to change the data before sending to
the server.

* Review Fixes

- Allow sending empty passwords to server
- If password is empty it will be silently dropped, so we loose
  the abilitty to unset passwords.

* Hide password input if password is set

As suggested by @bernd the password input field will be hidden
if a password is set. If the user would like to set a new password
he clicks on the new "Reset Password" button and the password
input field will be displayed.

* Add Review Suggestion from @bernd

As @bernd pointed out the Connection Test uses the form data
but with the password not hidden we need to take the password
from the database to test the connection.