Remove deprecated isMounted() calls
isMounted is marked as deprecated and raises warnings therefore.
The idea of the use of isMounted was to avoid setState() when
the component is already unmounted. That happens when setState()
is called in a promise.

Instead of using isMounted now I cancel the promise with the
help of bluebird.js in componentWillUnmount and check in the
then() challback if the promise was cancelled instead of
checking if the component is still mounted.