Remove warnings from browser console (#4605)
* Change LinkContainer to Link if child is a Anchor

LinkContainer is meant for React-Bootstrap components to
be used as navigation widgets with React-Router. But
using it with Anchors (<a>) will cause an error since
the prop active will be add to the component.
Instead of using LinkContainer with Anchor the
React-Router component Link should be used.

Also: I found a MenuItem not using LinkContainer
and fixed that.

* Fix warning with select and value null

A select component raises a warning when initilized with
null. Instead a empty string should be used.

* Fix warning of PropTypes

instanceOf(Immutable.List) did raise a warning that it expected
a List but got a List. Use already installed ImmutablePropTypes
to silent this warning.

* Remove deprecated isMounted() calls

isMounted is marked as deprecated and raises warnings therefore.
The idea of the use of isMounted was to avoid setState() when
the component is already unmounted. That happens when setState()
is called in a promise.

Instead of using isMounted now I cancel the promise with the
help of bluebird.js in componentWillUnmount and check in the
then() challback if the promise was cancelled instead of
checking if the component is still mounted.

* Remove all uses of instanceOf(Immutable)

Immutable does not go very well with PropTypes and created
warnings. To remove this warnings we use already installed
react-immutable-proptypes.