Separating web bindings for submodules. (Graylog2/graylog-plugin-enterprise#68)
Instead of one big binding file (`src/web/index.jsx`), we should have
bindings for each submodule of this plugin. This change is splitting up
the bindings file and groups them according to the submodule they
belong. The main bindings file is now collecting these and reduces them
to a binding which is registered.