Removing explicit inclusion of node type, leaving that to consumers.
With this change, the node type is not explicitly included in ts. It is
left to the components using node functions to include them. This allows
us to make it work for:

  a) compilations of projects with @types/node installed
  b) compilations of projects without @types/node.

For the latter, the inclusion of a reference path for the `WidgetsStore`
with a relative reference allows the compiler to find the @types/node
package of the graylog2-web-interface module.