Fix the way TypeScript loads type definitions. (#4642)
* Fix the way TypeScript loads type definitions.

Before this change, ts-loader/tsc just included the node types
explicitly. This lead to problems with duplicate inclusions for plugins
which transitively depend on typings.

This PR changes this behavior to the more standard one of defining which
types to include instead of including the types file explicitly.

* Removing explicit inclusion of node type, leaving that to consumers.

With this change, the node type is not explicitly included in ts. It is
left to the components using node functions to include them. This allows
us to make it work for:

  a) compilations of projects with @types/node installed
  b) compilations of projects without @types/node.

For the latter, the inclusion of a reference path for the `WidgetsStore`
with a relative reference allows the compiler to find the @types/node
package of the graylog2-web-interface module.

* Removing all type inclusion per default.