Add GraylogDataDir and GraylogBinPath config annotations
Configuration parameters that are annotated with GraylogDataDir and are
injected with the GraylogDataDir("<parameter-name>") qualifier use the
generic "data_dir" as a base directory if appropriate.

If the Path value of the config parameter is a relative path, the
"data_dir" setting will be used as base dir. If it's an absolute path it
doesn't use the "data_dir" as base dir.

This also introduces the "bin_path" and "data_dir" configuration
options.

Closes #2057