Adding keys to plugin routes.
During initialization of the `AppRouter`, plugin routes are added by
iterating over a collection. During this step, `Route` components are
generated without a `key`. This emits a warning generated by react and
is prone to collisions preventing individual routes to take effect.

This change is generating a key for each `Route` component consisting of
the plugin route component's `displayName` and its `path`. At least the
latter should be unique (otherwise one of the routes is shadowed
anyway).