Fixing code splitting for enterprise plugin. (Graylog2/graylog-plugin-enterprise#82)
Due to the current webpack config in graylog2-server, all components
that have a path matching `/pages\/.+\.jsx$/` will be splitted out into
separate chunks. Although this is a desired behavior, this is currently
not supported for plugins. Those chunks are not being picked up when the
plugin's assets are bundled into the jar file, because only the
entrypoints are referenced in the module manifest.

This PR moves the page component of the enterprise search one level
higher to avoid matching the aforementioned pattern and therefore
prevents being code split.