Fixing initial race condition between promises. (Graylog2/graylog-plugin-enterprise#83)
During the initial page load, stores (QueryStore & WidgetStore) are
hydrated with the initial query/widget structure. As this is performed
using the same actions as we do when manually adding queries/widgets,
multiple concurrent search executions are triggered.

Due to the behavior of the (old) version of reflux we are using, the
first search execution that succeeds (to be exact: the first creation of
a search object) is winning. At later stages, this search/result is used
against a more recent widget mapping structure contained in the search
request, which fails due to changed ids.

To fix this, we are now passing around the search request during search
creation, so it can be used later to map widgets against search types.