Add HOC generator connecting ES6 class components to stores.
This generator allows ES6 class components to be connected to stores,
although the current version of Reflux allows this only using a mixin,
which is not usable for ES6 classes.

The change provides a `connect()` function which takes the supplied
component class and wraps it in a generated component which connects
itself to the stores defined in the provided mapping. The state supplied
by the stores is then passed to the component as props.