Avoid DOM structure changes when there are no pages to show
This fixes a long standing issue with the SearchForm component where the
query string disappears when a search doesn't return any results.

The problem was that we had a shortcut in the render function that
returned only the children and no page selectors when there are no items
to show. This apparently changed the DOM structure in a way that the
children of the PaginatedList component have been replaced with new
instances.
In the case of a SearchForm component as a PaginatedList child, this
reset the query string.