Reinitialize view when starting new search.
Before this change, a loaded view would still be present when the
"Extended Search" navigation point or "Create new view" button was used
to start a new search.

This change reinitializes the view in a new entry point (NewSearchPage
component) which is used only for interactions that start a new search.