Measuring width before mounting widget grid.
The widget grid is now measuring the container width before mounting the
component and rendering widgets. This way accidental resizing of
oversized widgets or widgets with a high x coordinate on container
widths higher than 1280 (which is the default for the initial rendering
when `measureBeforeMount` is `false`).

This change is cloning the `ReactGridContainer` component and its npm
dependencies (`react-grid-layout` and `react-resizable`) until
Graylog2/graylog2-server#4723 is merged.