Adapt default search operator to type
Using `SearchQueryOperators.REGEX` as default is fine for String types,
but it forces users to type an operator on Date and Int types.

This commit leaves `REGEX` as default for Strings, but uses `EQUALS` for
Dates and Ints, so users can still find values if they only write the
value.