Add prop to render help trigger in SearchForm (#4734)
`SearchForm` is almost always used with some special query syntax that
that users may not know, or may not always remember. That means that
almost always there is a help button associated to the `SearchForm`. So
far the help button was rendered outside the `SearchForm` component,
usually too far away from the text input as to still remain some
context.

This commit let us render a help button inside the text input, helping
users to associate that help with the text that they are supposed to
write in that input.