Replacing babel es2015 preset with env preset. (#4749)
The babel team is recommending to switch to the `env` preset instead of
using the yearly presets (which also require continuous updates to the
config). According to the babel team the preset has: `By default, [...]
the same behavior as previous presets to compile ES2015+ to ES5.`

More information can be found [here](https://babeljs.io/env/).