Initializing map holding element refs in `GeoIpResolverConfig`.
In #4736, string refs have been replaced by callbacks. As a fallout of
this change, the map holding refs for inputs in `GeoIpResolverConfig`
was forgotten to initialize, breaking the Geo Ip Processor config modal.

This change is fixing this. Additionally, #4736 was checked again for
variables holding element refs in a nested way (like
`this.inputs[element]`) but no other occurrences where the holder was
not initialized were found.