Add `trigger_alert()` function
Alert conditions in Graylog are evaluated in a fixed interval by
`AlertScanner` and `AlertScannerThread`, which might be too inflexible
for some users who want to trigger an immediate alert if a specific
message is encountered.

In order to enable users to manually trigger alerts in a pipeline rule,
this change set adds the `trigger_alert()` function.

The function also enables a primitive kind of "real-time alerting"
utilizing pipeline rules.