Add stream ID to stream rule metrics name
The original metrics name didn't contain any information about the stream which the stream rule was part of.

This made it very hard to identify the stream which included the stream rule and was only possible by either
querying MongoDB directly or take a look into each stream (brute force).

This change set changes the metrics name from

    org.graylog2.plugin.streams.StreamRule.${stream-rule-id}.executionTime

to

    org.graylog2.plugin.streams.Stream.${stream-id}.StreamRule.${stream-rule-id}.executionTime

Refs https://community.graylog.org/t/identify-stream-rule-from-metrics/5266