Do not update positions after click in widget actions
Clicking anywhere on an unlocked widget generates a drag event, which in
our case ends up sending a request to the server to update widgets
positions.
Use React Grid Layout `draggableCancel` prop, to avoid calling drag
callbacks when a user clicks on the edit or delete buttons for a widget.

Partial backport of 9488608 (Graylog2/graylog2-server#4530).
Mitigates #4525