Using FieldTypeMapper to map physical types onto logical types. (Graylog2/graylog-plugin-enterprise#214)
* Using FieldTypeMapper to map physical types onto logical types.

This change employs the `FieldTypeMapper` class in the
`FieldTypesResource` to map ES types to Graylog types before returning
field types. The frontend code is adapted to the structure which is
returned. As an example, the "Chart" action eligibility check is changed
from checking the field name to checking for the `numeric` property in
the field type.

* Object is not an immutable map anymore.

* Updating yarn.lock.