Preventing rerendering of Search Result if query changes.
Before this change, upon every change of the query string (read: every
single character typed in the query input), the search result
(represented by the `SearchResult` component) was rendered again, as it
is subscribed to the `SearchStore`. This made the typing experience in
the `QueryInput` component unbearably sluggish.

With this change, we are doing two things:

  1) Strip search object from state of `SearchStore` in `SearchResult`
by using the props mangling function passed to `connect()`
  2) Turn the `SearchResult` into a class-based component again and
implement a `shouldComponentUpdate` function doing a deep equality check
of props (which takes significantly less time than rerendering the whole
subtree) preventing rerendering if no props have changed.