Portalize widget action dropdown menu. (Graylog2/graylog-plugin-enterprise#232)
Before this change, the widget action dropdown menu (triggered by the
caret on the upper right of a widget) could be cut off if the widget
size was too small.

After this change, the widget action dropdown menu is portalized
(positioned *outside* of the container of the widget which has `overflow`
set to `hidden) and therefore is not cut off.

The positioning of the dropdown is a bit off now. It needs to be
positioned slightly to the right and to the bottom. This requires some
fiddling with the internals of the `Overlay` component from
`react-overlays`, so it is saved for a future PR.