Fixing widget store state propagation. (Graylog2/graylog-plugin-enterprise#230)
In Graylog2/graylog-plugin-enterprise#208 stores were centralized, so store states are propagated from a
central store down to subscribers of specific parts. In this change, the
widget store was changed to get its state from a parent store, but still
manipulated its own internal state on actions. Due to the check in the
parent store update handler for equality of old and new state, the new
state received was never propagated to subscribers of the store.

This change now leaves setting the internal state of the `WidgetStore`
to the parent store update handling function only.