Add "Add to query" action in field value action menu. (Graylog2/graylog-plugin-enterprise#226)
This feature is reimplementing the functionality of the looking glass
button to the right of field values in message details. Instead of a
separate button this is integrated into the dropdown menu activated by
clicking the field's value.

When the search bar is out of sight, a hint is displayed that something
was changing the search bar, just as on the current search page.

Fixes Graylog2/graylog-plugin-enterprise#156.