Reimplement streams functionality in message details. (Graylog2/graylog-plugin-enterprise#224)
* Reimplement streams functionality in message details.

Before this change, the list of streams a message was routed to and the
dropdown for testing a message against a stream was empty in the new
extended search.

This change is reimplementing the functionality. For this, it creates a
Reflux store named `StreamsStore`, which is wrapping the existing TS
`StreamsStore`. The latter only provided actions to retrieve and modify
streams, but did not keep state, requiring round trips to the server for
every component using it. The new `StreamsStore` is doing this once upon
initialization and memoizes the result.

It is desirable in the future to replace the existing `StreamsStore`
completely, but due to the amount of components relying on its current
semantics, this is way beyond the scope of this PR.

Fixes Graylog2/graylog-plugin-enterprise#160.

* Removing now obsolete references.