Reenable buttonAfter props for Input component.
Since #3598 / bba0e6beb7cff91307e4d9e412314dac825a9425 components which
are passed via the `buttonAfter` prop to the `Input` component are not
rendered anymore, due to changes of the react-bootstrap API. This leads
to inaccessible functionality as described in e.g. #4815.

This PR is reestablishing the used functionality.

Fixes #4815.