Add option to render Wizard in horizontal (#4853)
Sometimes we want to provide a wizard having some extra width for rendering components, e.g. when rendering a wizard inside a modal. In these cases, it is good rendering our `Wizard` component in horizontal, allowing us to save some space. This PR adds support for that.

Additionally, I also added a change, making the `content` CSS class in the outer row of the `Wizard` component customisable, as we may not always want to draw a border around the wizard.