Use backend support for widget filters. (Graylog2/graylog-plugin-enterprise#280)
* Add proper babel presets and module mapping for Jest.

This change adds the following for Jest tests:

  * `stage-0` and `flow` presets to babel config
  * a module mapping so imports starting with `enterprise` are properly
resolved

* Use backend support for widget filters.

This change extracts `filter` definitions (added by the widget filter
control in the UI) and adds them to the generated search types
definition when a new search is generated. This enables filtering of
widget data for widgets other than the message list.

Requires Graylog2/graylog-plugin-enterprise#279 to be merged before.

Fixes Graylog2/graylog-plugin-enterprise#170.