Allow values to be `null` for values in pivot results. (Graylog2/graylog-plugin-enterprise#285)
This change allows values in pivot results to be `null`. This happens,
when the field is not included in any message of the bucket that is
being aggregated. Instead of failing the complete search, this change
allows those values to be `null`, so they can be handled properly by the
client instead.