Requires user to be authenticated to retrieve plugin list. (#4868) (#4875)
Before this change, the `SystemPluginResource` which returns the list of
installed plugins for this node, did not require any authentication at
all. This might lead to unnecessary disclosure of harmful information
and should be avoided.

Therefore this change adds the annotation which requires the caller of
the `SystemPluginResource` to be authenticated. If this is not
sufficient, a further check for a permission can be introduced.

Fixes #4863.

(cherry picked from commit 267be4ae23b3ef163427bbc816dd2c7378f755e4)