Use proper aggregation type classes for chart action. (Graylog2/graylog-plugin-enterprise#278)
Before this change, the "Chart" field action handler used the old,
untyped widget definition to create a chart widget for the field. This
broke, when you tried to edit the widget.

This change uses the proper classes for the aggregation widget and its
config, which allows the aggregation builder to read it back.