Use field filtering function/list from core instead of duplicating it. (Graylog2/graylog-plugin-enterprise#289)
While working on Graylog2/graylog-plugin-enterprise#284/#263, I accidentally duplicated functionality
(namely the `MessageFormatter` function), which is already present in
core. This change is removing the duplicated functionality and
references to core.

Additionally, it fixes a bug introduced in Graylog2/graylog-plugin-enterprise#263, which leads to no field
starting with `gl2_` be shown in the sidebar per default, although only
specific ones whould be filtered out (as defined in the
`FILTERED_FIELDS` list of the `MessageFieldsFilter` module.