Regenerate Search id upon structural changes to search. (Graylog2/graylog-plugin-enterprise#294)
Before this change, the same search id was over and over again for a
given view. This lead to the problems that some parts of the view (query
string, time ranges) where persisted not when a view is saved, but when
a search using these changed properties was performed, even if the view
as not saved afterwards.
This happened because the search id is associated with the view and did
not change. The reason for reusing the search id was trying to prevent
pollution of the `searches` collection when no cleanup job was in place.

This change is now regenerating the id of the `Search` object for every
structural change. Obviously this now leads to a much higher (and
possibly boundless) number of documents in the `searches` collection.
Therefore a followup PR will be a periodic cleanup job of that
collection.