Support infinity values for widget positions (Graylog2/graylog-plugin-enterprise#304)
* Using custom class for widget positions to allow values of Infinity.

Before this change, widget positions were kept as numbers/ints in
front-/backend. Due to the lack of support for serializing `Infinity` in
JSON, it was not possible to specify this for any widget position value.

It is desirable to be able to use `Infinity` e.g. for the width of a
widget, so it takes up all available vertical screen space. Therefore
this PR adds support for values of `Infinity`, by wrapping widget
positions in separate classes with custom (de-)serialization.

* Add stretch button to widget

Add a button to stretch a widget horizontally. The icon of the button will
change to compress so a user can bring it back to defautltWidth

* Fix annotations by @dennis