Check if user is allowed to see streams in Search. (Graylog2/graylog-plugin-enterprise#305)
Before this change, there was no permission enforment when performing
searches. Additionally, the indices a query was performed in was
limited by the given time range only, but not the streams it is
searching in.

This change is now doing several things:

  - Add method to `Query` class which returns referenced stream ids
  - Check if user is allowed to read all stream ids referenced in all
queries of a `Search` before it is executed
  - If no stream filters are given for a `Query`, the query is
implicitly scoped to all streams the calling user has access to
  - Only indices are used for the search which:
    - Contain the given time range
    - Contain at least one of the referenced streams
  - If the set of indices used for the execution of a query is empty, it
is not falling back to use `_all`, but use nothing instead.

Refs Graylog2/graylog-plugin-enterprise#167.
Fixes Graylog2/graylog-plugin-enterprise#241.