Add "http_non_proxy_hosts" configuration option (#4908)
This can be used to bypass a configured proxy server for a list of
hostnames or IP addresses. It must be used if the inter-node
communication between graylog servers should not go through a configured
proxy server.

The matcher implementation for the "http_non_proxy_hosts" is similar to
the implementation that is used for the "http.nonProxyHosts" system
property. (only the delimiter is "," instead of "|" for consistency)
By using a similar implementation, it will be possible to also set the
system property in the future. (if needed - tbd)

Fixes #4905
Fixes #4392

(This needs to be cherry-picked into 2.4 once merged)