If `activeQuery` is not in list of queries, use first one when loading. (Graylog2/graylog-plugin-enterprise#314)
* If `activeQuery` is not in list of queries, use first one when loading.

Before this change, when loading a view (either by loading a fresh,
persisted one or by reloading the current one which was saved), it was
checked if there is a current active query. If yes, this query was
selected instead of the first query. When a view was "switched" though,
there was still an active query, but this one was most certainly not
present in the view that was switched to, therefore no active query was
selected.

This change implements a check if the active query is present in the
list of queries and if not, falls back to selecting the first query.

Fixes Graylog2/graylog-plugin-enterprise#311.

* Adding tests to cover issue and changes.

* Remove unused reference.