Show spinner when loading widget data. (Graylog2/graylog-plugin-enterprise#310)
* Support testing react components in jest.

This change adds the following features:

  * Adding required packages/config for enzyme shallow/mount
  * Adding required package for react test renderer
  * Add module lookup for core components
  * Add module lookup for core test helpers

* Show spinner when loading widget data.

This change implements showing a spinner when widget data is not present
yet. This is not only for aesthetic purposes, but also fixes a bug where
the position of a newly created widget was already part of the layout,
but the widget itself was not rendered yet due to the check for
widget data presence in `WidgetGrid`. The consequence of this was that
the newly created widget had invalid dimensions.

* Adding tests.

* Fix build after broken merge conflict resolution