Include `count()` in function autocompletion. (Graylog2/graylog-plugin-enterprise#325)
This PR is statically adding the `count()` function to the
autocompletion in the series controls. While all other functions are
retrieved dynamically from the backend on runtime, this is wired in
statically for now. We are currently lacking the ability to provide
parameters in the series descriptors returned by the backend. This is
the scope of future enhancements.