Fix the dialog shown to the user when leaving page and view is dirty. (Graylog2/graylog-plugin-enterprise#327)
This PR is changing the check if the current view is dirty when a user
is leaving the page (which might result from clicking the back button,
entering a different URL or reloading the page). Due to restructuring
the `ExtendedSearchPage` component for fewer rerenderings, the dirty
flag got lost. While fixing this, the logic for this was extracted into
a separate component and a check was implemented to disable the
functionality if development mode is active.