Prevent value deduplication for different parent values in Data Table. (Graylog2/graylog-plugin-enterprise#344)
Before this change, a children k/v pair was deduplicated (value was
removed) when the last value in this column was identical. This is
misleading if any of the parent values changed.

This change removes children's values from the duplicates list if a
value changes and therefore prevents this.

For better testing, the function doing the deduplication was extracted
from the `DataTable` component and tests were added.