Fix deserialization of legacy single string timeunit intervals. (Graylog2/graylog-plugin-enterprise#365)
This is a quick fix for Graylog2/graylog-plugin-enterprise#358. Deserialization behavior for Interval
instances changed, so that instead of defaulting to a time unit which is
described by a single string, a struct with a `type` field is expected,
causing attempts to deserialize `Search` objects created by pre-#358
code to fail.

This change enables parsing of single string `Interval` instances,
defaulting to the `TimeUnitInterval` implementation and treating the
single string as `timeunit` declaration.