Add validation to mandatory fields and better arrange overview
A user new to the content packs might not understand what he
has to fill out to get to the next page. Now we show which fields
a necessay to complete the general information and content pack selection.

The preview of content pack edit/creation was wasting space.
Now all list are on the right side and all general information
are placed to the left.