Add new StoreProviderMock and add custom Function to StoreMock (#4914)
* Add new StoreProviderMock and add custom Function to StoreMock

Add a new StoreProviderMock which can be used as a mock
for StoreProvider. It's a factory which takes a map of stores
with the store names as keys.

Also enhance StoreMock so it can add a custom function
in stead of jest.fn(). This is useful when the function
should return something specific or you want to add a
`expect` to the function outcome.

* Fix annotations from @dennisoelkers