Allow conditions for value actions. (Graylog2/graylog-plugin-enterprise#376)
When clicking a value rendered using the `Value` component (e.g. a
value in a data table), an action dropdown menu is shown which allows
further interaction to the user.
This change implements conditional rendering of value action based on
the field name, type and value of the clicked value. When binding a
value action, a new key `condition` can be used, which expects a
function returning a value, that is interpreted/coerced to a boolean
indicating if the value action is disabled or not.