Introduce permissions for Views/Extended Search features. (Graylog2/graylog-plugin-enterprise#381)
This change set is adding basic permission defining if a user is allowed
to use the views/extended search features or not. For this, the
`view:use` and `extendedsearch:use` permissions are introduced.
Furthermore, a migration creating a built-in role including these
permissions and the permissions to create views/extended searches was
added.

Requires Graylog2/graylog2-server#5003 to be merged before.