Avoid series name collisions when n>1 search types have `rollup` set. (Graylog2/graylog-plugin-enterprise#384)
* Avoid series name collisions when n>1 search types have `rollup` set.

This change is fixing an issue that occurs when there is more than one
`pivot` search type with `rollup` set to `true`. In those cases series
are added in the global scope outside of the aggregation as well, what
leads to a name collision as series are named as `series-<number of
series in aggregation>`.
A past attempt to solve this (Graylog2/graylog-plugin-enterprise#318) prefixed series names with a new
random identifier when row/column pivot lists are empty for the
aggregation. This is not sufficient, as the collision also occurs when
pivots exist, as the global rollup series is added outside of the scope
of the aggregation.

This change is now prefixing series names with the pivot id to generate
unique series names outside of the scope of the aggregation as well.

Fixes Graylog2/graylog-plugin-enterprise#382.

* Removing now unused `idPrefix` assignment.