Return default timezone if user timezone is not available. (Graylog2/graylog-plugin-enterprise#390)
Under some circumstances, there is a race condition between fetching the
user (and the timezone) and displaying the search result overview. In
this case an exception was thrown, as `currentUser` was `null`.

With this change, timezone extraction is performed using `get`, which
returns a default value (`UTC` in this case) if one of the components
being accessed is `null` or `undefined`.